revert last change, add an init container
An init container is created which will run raw commands specified
as string value to init: key in the values file.

The init container will by default enabled IP forwarding, disable eth0
and add an iptables rule to correct the UDP checksum for the packets
in transit to avoid errors with the checksum offloading feature.